--- src/App.css
@@ -1,8 +1,0 @@
-@import url('./blocks/Global.css');
-@import url('./blocks/HomePage.css');
-@import url('../src/vendor/fonts.css');
-@import url('../src/blocks/SearchBar.css');
-@import url('./blocks/Cards.css');
-@import url('./blocks/CarouselSection.css');
-@import url('./blocks/Header.css');
-@import url('./blocks/Footer.css');
