--- src/blocks/Carousel.css
@@ -1,14 +1,0 @@
-.carousel{
-    max-width: 1200px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.slick-arrow{
- background-color: black; 
-}
-
-.slick-arrow:hover{
-    background-color: black;
-    opacity: .8;
-}
