+++ src/components/CarouselSection/Cards.jsx
@@ -1,0 +1,12 @@
+import "./Cards.css";
+const Cards = ({ item }) => {
+  return (
+    <div className="card">
+      <img src={item.link} alt={item.name} className="card__image" />
+      <h3 className="card__name">{item.name}</h3>
+      <p className="card__description">{item.description}</p>
+    </div>
+  );
+};
+
+export default Cards;
